requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.1
djangorestframework==3.15.1
idna==3.7
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
kenar==0.4.3
pydantic==2.7.4
httpx==0.27.0